refactor(renderer): give the renderer its own helpers and logging
