+++ src/modules/ui/lyricsHost.ts
@@ -1,9 +1,10 @@
-import { TAB_HEADER_CLASS, TAB_RENDERER_SELECTOR } from "@constants";
+import { LOG_PREFIX, TAB_HEADER_CLASS, TAB_RENDERER_SELECTOR } from "@constants";
 import { t } from "@core/i18n";
 import { seekPlayer } from "@modules/lyrics/lyrics";
 import { hideAdOverlay, isAdPlaying, isLoaderActive, showAdOverlay } from "@modules/ui/dom";
 import { getResumeScrollElement } from "@modules/ui/resumeScrollButton";
 import type { LyricsRendererHost } from "@renderer/index";
+import { log } from "@utils";
 import { resetDebugRender, resizeCanvas } from "./animationEngineDebug";
 
 const PLAYER_PAGE_ID = "player-page";
@@ -52,5 +53,11 @@ export const ytmHost: LyricsRendererHost = {
   },
   seek: seekPlayer,
   translate: t,
+  /**
+   * Resolved per call rather than bound once: `log` is reassigned when the logging setting loads.
+   */
+  log(...args: unknown[]): void {
+    log(LOG_PREFIX, ...args);
+  },
   debug: { beginFrame: resetDebugRender, resize: resizeCanvas },
 };
